Moved validators from cadc-tap-schema to cadc-tap. Commented ucd files related tests.
