+++ cadc-tap/src/main/java/ca/nrc/cadc/tap/schema/validator/Violation.java
